config/rdt: separate status for "enabled" and "supported"
Rename Enabled() to Supported() - it tells if RDT is supported (and
enabled) in the underlying host HW/OS.

Add a new Enabled() method for tracking if RDT has been enabled in
cri-o. We want to track this for e.g. cases where a configuration update
(reload) effectively disables RDT by not specifying any RDT
configuration file.